jae-scanner-trigger/backend/requirements.txt
Flask
python-dotenv
google-generativeai
pillow
gunicorn
google-cloud-storage
cloudevents
functions-framework